docs: Use a clear name for the Polarion template
Before, it was not clear which template to select in the list of templates.
The template name or description didn't contain any reference to Capella or
Capella2Polarion.

In addition, it had a reference to "Deutsche Bahn", but the template can
of cause also be used outside of Deutsche Bahn.